# produtos-api/docs/swagger.yaml
definitions:
  models.Product:
    description: A product model
    properties:
      description:
        description: Product Description
        type: string
      id:
        description: Product ID
        type: integer
      name:
        description: Product Name
        type: string
      price:
        description: Product Price
        type: number
      stock:
        description: Product Stock
        type: integer
    type: object
info:
  contact: {}
paths:
  /products:
    get:
      consumes:
      - application/json
      description: Retorna todos os produtos
      parameters:
      - description: Nome do produto
        in: query
        name: name
        type: string
      - description: Contagem de produtos
        in: query
        name: count
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Retorna todos os produtos
      tags:
      - produtos
    post:
      consumes:
      - application/json
      description: Cria um novo produto
      parameters:
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Cria um novo produto
      tags:
      - produtos
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Deleta um produto
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Deleta um produto
      tags:
      - produtos
    get:
      consumes:
      - application/json
      description: Retorna um produto pelo ID
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Retorna um produto pelo ID
      tags:
      - produtos
    put:
      consumes:
      - application/json
      description: Atualiza um produto
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: integer
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Atualiza um produto
      tags:
      - produtos
swagger: "2.0"